Projects/Hex_Card_Forge/cardforge/tests/test_image.py
"""
Tests for image processing functionality
"""

import os
import tempfile
from pathlib import Path
import pytest
from PIL import Image, ImageChops

from cardforge.image import crop_hex


def create_test_image(size=1024, color='white'):
    """Create a square test image with a solid color."""
    img = Image.new('RGBA', (size, size), color=color)
    return img


def count_transparent_corners(img, threshold=20):
    """Count the number of transparent corners in an image.
    
    Args:
        img: PIL Image in RGBA mode
        threshold: Pixel distance from corner to check for transparency
        
    Returns:
        int: Number of transparent corners (0-4)
    """
    width, height = img.size
    corners = [
        (0, 0),  # top-left
        (width-1, 0),  # top-right
        (0, height-1),  # bottom-left
        (width-1, height-1)  # bottom-right
    ]
    
    transparent_corners = 0
    for x, y in corners:
        # Check if corner pixel is transparent
        if img.getpixel((x, y))[3] == 0:
            transparent_corners += 1
    
    return transparent_corners


def test_crop_hex_basic():
    """Test that crop_hex creates a proper hexagon with transparent corners."""
    # Create a temp directory for our test images
    with tempfile.TemporaryDirectory() as temp_dir:
        # Create a test square image
        square_path = os.path.join(temp_dir, "square.png")
        square_img = create_test_image(size=1024, color='red')
        square_img.save(square_path)
        
        # Crop to hexagon
        hex_path = crop_hex(square_path)
        
        # Verify the result exists
        assert os.path.exists(hex_path)
        
        # Load the resulting image
        hex_img = Image.open(hex_path)
        
        # Verify dimensions
        assert hex_img.width == 1024, f"Expected width 1024, got {hex_img.width}"
        assert hex_img.height == 1024, f"Expected height 1024, got {hex_img.height}"
        
        # Verify it has an alpha channel
        assert hex_img.mode == "RGBA", f"Expected RGBA mode, got {hex_img.mode}"
        
        # Check for transparent corners (flat-top hex should have 6 corners with alpha=0)
        transparent_count = count_transparent_corners(hex_img)
        assert transparent_count == 4, f"Expected 4 transparent corners, found {transparent_count}"
        
        # Calculate the area of non-transparent pixels
        non_transparent = sum(1 for x in range(hex_img.width) for y in range(hex_img.height) 
                             if hex_img.getpixel((x, y))[3] > 0)
        
        # For a regular flat-top hexagon with width w, the area is 3√3/2 × (w/2)²
        # Which simplifies to 0.75 × √3 × w² or approximately 0.75 × 1.732 × w²
        # This is roughly 0.65 × w² for a regular hexagon with side length = width/2
        # However, since we're fitting the hexagon inside a square, the area is less than that
        expected_area = int(0.65 * 1024 * 1024)  # Updated to more accurate value
        area_tolerance = 0.15  # Increased tolerance to 15% due to pixel approximation
        
        assert abs(non_transparent - expected_area) < expected_area * area_tolerance, \
            f"Hexagon area {non_transparent} differs from expected {expected_area} by more than {area_tolerance*100}%"
