Projects/Hex_Card_Forge/requirements.txt
typer>=0.9.0
pillow>=10.0.0
pytest>=7.3.1
rich>=13.3.5
